arm64 kexec: separate testable code
